Update lambda runtime to nodejs12.x
Remove the hardcoded runtime environment in the module. This can be
specified by the caller, allowing the runtime to be upgraded more easily.

Computed the millisecond equivalent of 6 minutes incorrectly at 6000.
That's 6 *seconds*. We want 6 * 60 * 100, which comes to 360000. Oops.

Fixed a bunch of minor documentation errors.

Added support for tags.

While we're at it, got rid of Drone.